+++ app/src/main/java/com/alvindizon/tampisaw/di/app/RepoModule.kt
@@ -1,19 +1,35 @@
 package com.alvindizon.tampisaw.di.app
 
-import com.alvindizon.tampisaw.api.UnsplashApi
-import com.alvindizon.tampisaw.data.UnsplashRepoImpl
-import com.alvindizon.tampisaw.domain.UnsplashRepo
+import com.alvindizon.tampisaw.collections.integration.CollectionsViewRepository
+import com.alvindizon.tampisaw.details.integration.DetailsViewRepository
+import com.alvindizon.tampisaw.gallery.integration.GalleryViewRepository
+import com.alvindizon.tampisaw.integration.CollectionsViewRepositoryImpl
+import com.alvindizon.tampisaw.integration.DetailsViewRepositoryImpl
+import com.alvindizon.tampisaw.integration.GalleryViewRepositoryImpl
+import com.alvindizon.tampisaw.integration.SearchViewRepositoryImpl
+import com.alvindizon.tampisaw.repo.UnsplashRepo
+import com.alvindizon.tampisaw.repo.UnsplashRepoImpl
+import com.alvindizon.tampisaw.search.integration.SearchViewRepository
+import dagger.Binds
 import dagger.Module
-import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
-class RepoModule {
+abstract class RepoModule {
+    @Binds
+    abstract fun provideUnsplashRepo(repo: UnsplashRepoImpl): UnsplashRepo
 
-    @Provides
-    @Singleton
-    fun provideUnsplashRepo(unsplashApi: UnsplashApi): UnsplashRepo = UnsplashRepoImpl(unsplashApi)
+    @Binds
+    abstract fun provideCollectionsViewRepo(repo: CollectionsViewRepositoryImpl): CollectionsViewRepository
+
+    @Binds
+    abstract fun provideDetailsViewRepo(repo: DetailsViewRepositoryImpl): DetailsViewRepository
+
+    @Binds
+    abstract fun provideGalleryViewRepo(repo: GalleryViewRepositoryImpl): GalleryViewRepository
+
+    @Binds
+    abstract fun provideSearchViewRepo(repo: SearchViewRepositoryImpl): SearchViewRepository
 }
