+++ app/src/main/java/com/alvindizon/tampisaw/integration/CollectionsViewRepositoryImpl.kt
@@ -6,11 +6,14 @@ import com.alvindizon.tampisaw.collections.integration.CollectionsViewRepository
 import com.alvindizon.tampisaw.collections.model.Collection
 import com.alvindizon.tampisaw.core.toCollection
 import com.alvindizon.tampisaw.core.toPhoto
-import com.alvindizon.tampisaw.domain.UnsplashRepo
 import com.alvindizon.tampisaw.gallery.model.Photo
+import com.alvindizon.tampisaw.repo.UnsplashRepo
 import io.reactivex.rxjava3.core.Observable
+import javax.inject.Inject
+import javax.inject.Singleton
 
-class CollectionsViewRepositoryImpl(private val unsplashRepo: UnsplashRepo) :
+@Singleton
+class CollectionsViewRepositoryImpl @Inject constructor(private val unsplashRepo: UnsplashRepo) :
     CollectionsViewRepository {
 
     override fun getAllCollections(): Observable<PagingData<Collection>> {
