+++ app/src/main/java/com/alvindizon/tampisaw/integration/DetailsViewRepositoryImpl.kt
@@ -3,10 +3,14 @@ package com.alvindizon.tampisaw.integration
 import com.alvindizon.tampisaw.core.toPhotoDetails
 import com.alvindizon.tampisaw.details.integration.DetailsViewRepository
 import com.alvindizon.tampisaw.details.model.PhotoDetails
-import com.alvindizon.tampisaw.domain.UnsplashRepo
+import com.alvindizon.tampisaw.repo.UnsplashRepo
 import io.reactivex.rxjava3.core.Single
+import javax.inject.Inject
+import javax.inject.Singleton
 
-class DetailsViewRepositoryImpl(private val unsplashRepo: UnsplashRepo) : DetailsViewRepository {
+@Singleton
+class DetailsViewRepositoryImpl @Inject constructor(private val unsplashRepo: UnsplashRepo) :
+    DetailsViewRepository {
 
     override fun getPhotoDetails(id: String): Single<PhotoDetails> {
         return unsplashRepo.getPhoto(id).map {
